+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/components/settings/Settings.tsx
@@ -7,7 +7,7 @@ import ClayEmptyState from '@clayui/empty-state';
 import React from 'react';
 
 import {getImage} from '../../../common/utils/getImage';
-import useSelectedItem from '../../contexts/hooks/useSelectedItem';
+import useSelectedItem from '../../hooks/useSelectedItem';
 import ReferencedStructureSettings from './ReferencedStructureSettings';
 import RepeatableGroupSettings from './RepeatableGroupSettings';
 import StructureFieldSettings from './StructureFieldSettings';
