+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/hooks/useSelectedItem.ts
@@ -3,16 +3,16 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import selectSelection from '../../selectors/selectSelection';
-import selectStructureChildren from '../../selectors/selectStructureChildren';
+import {useSelector} from '../contexts/StateContext';
+import selectSelection from '../selectors/selectSelection';
+import selectStructureChildren from '../selectors/selectStructureChildren';
 import {
 	ReferencedStructure,
 	RepeatableGroup,
 	Structure,
-} from '../../types/Structure';
-import {Uuid} from '../../types/Uuid';
-import {Field} from '../../utils/field';
-import {useSelector} from '../StateContext';
+} from '../types/Structure';
+import {Uuid} from '../types/Uuid';
+import {Field} from '../utils/field';
 
 type SelectedChild =
 	| {field: Field; referenced: boolean; type: 'field'}
